requirements.txt
psycopg2-binary
django
djangorestframework
django-cors-headers
gunicorn
django-rest-knox
debugpy
